pyproject.toml
[tool.poetry]
name = "vacation-manager"
version = "0.1.0"
description = ""
authors = ["Dmitry <erofeevdma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
python-dotenv = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
